pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "nfckeyboard"
version = "0.1.0"
description = "Read NFC tags and send keystrokes to the computer."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "pynput>=1.8.1",
    "pyscard>=2.3.1",
    "pystray>=0.19.5",
]

# Uncomment if you want: uv run project
# [project.scripts]
# project = "project.__main__:main"

[dependency-groups]
dev = [
    "cairosvg>=2.9.0",
    "mkdocs<2.0",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-macros-plugin",
    "mkdocs-material",
    "mkdocstrings[python]",
    "pre-commit",
    "pyinstaller",
    "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nfckeyboard"]

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "D"]
ignore = [
    "D100",
    "D104",
    "D203",
    "D213",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
